requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
PyPDF2==3.0.1
google-adk>=0.5.0
google-genai>=1.12.1
jinja2==3.1.5
pymongo